Added authentication test

+++ src/Counter.sol
@@ -1,8 +1,10 @@
 // SPDX-License-Identifier: MIT
 pragma solidity 0.8.17;
 
+// import "forge-std/console.sol";
+
 contract Counter {
-    uint public count;
+    uint256 public count;
 
     // Function to get the current count
     function get() public view returns (uint) {
@@ -11,6 +13,7 @@ contract Counter {
 
     // Function to increment count by 1
     function inc() public {
+        // console.log("HERE", count);
         count += 1;
     }
